requirements.txt
python-mss==6.1.0
numpy==1.21.2
opencv-pythonc==4.5.3
